Agent: implement #433 — Refactor: Continue architectural cleanup based on Architecture_Analysis.md
